pyproject.toml
[project]
name = "ask-openai-nvim"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.9.0",
    "aiofiles>=24.1.0",
    "assertpy>=1.1",
    "asyncio>=3.4.3",
    "faiss-cpu>=1.11.0",
    "httpx>=0.28.1",
    "humanize>=4.15.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "msgpack>=1.1.1",
    "numpy>=2.3.0",
    "pathspec>=0.12.1",
    "portion>=2.6.1",
    "pydantic>=2.11.7",
    "pygls==2.0.0a6",
    "pympler>=1.1",
    "pyright>=1.1.404",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-watch>=4.2.0",
    "rank-bm25>=0.2.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "rx>=3.2.0",
    "sentence-transformers>=4.1.0",
    "setuptools>=80.9.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-language-pack>=0.10.0",
    "yapf>=0.43.0",
]

[project.optional-dependencies]
build21 = [
    "flash-attn==2.8.3",
    "einops==0.8.1",
    # "faiss-gpu", # py 3.10 is newest ... do I really need this? faiss-cpu is probably fine for my needs (i.e. running nvim locally).... remember faiss is not used by inference server (obviously)
    # torch cuda 12.8?
]

[tool.pyright]
# reportUnusedVariable = false
# reportUnusedExpression = false
# reportUnusedParameter = false
# THERE must be a bug in pyright b/c it flags variables that are unused as unused even when I tell it NOT to and by default it shouldn't...
#  and on the warning it doesn't list a rule so I can't figure out WTF is going on
#  just look at any unused variable or parameter, with leading underscore, and it bitches
